rename algorithm_tests to algorithm_tests_without_privacy, move file lib.py  to tests folder
