--- Exareme-Docker/src/mip-algorithms/tests/algorithm_tests/lib.py
@@ -1,1 +1,0 @@
-vmUrl = 'http://localhost:9090/mining/query/'
