+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests_without_privacy/test_pca.py
@@ -4,7 +4,7 @@ import requests
 import math
 import numpy as np
 
-from lib import vmUrl
+from tests.lib import vmUrl
 
 endpointUrl = vmUrl + 'PCA'
