+++ Exareme-Docker/src/mip-algorithms/tests/lib.py
@@ -1,0 +1,1 @@
+vmUrl = 'http://88.197.53.34:9090/mining/query/'
